Item11267: No,, we need to use version 0.77
It seems the only sane way to get to modern perl version strings that
are usable, printable, comparable,  is to require minimum version 0.77

Before version 0.77,  "v" style strings cannot be quoted to be useful as
versions, but cannot be printed unless quoted.

Version 0.77 shipped with Perl 5.10.1,  Older versions that must be
upgraded shipped starting with Perl 5.9.0.

git-svn-id: http://svn.foswiki.org/trunk/HomePagePlugin@16013 0b4bb1d4-4e5a-0410-9cc4-b2b747904278